Minor cleanup. Added IrWriter. Began working on file format.
Did some minor cleanup and refactoring.
Added ability to remove attachments.
Removed the UsageAnalysis attachment after it was no longer needed/valid.
Added IrWriter.
Added attachment for replacing the IR code of a function.
Added handling of SSA branches to the SsaVisitor.
Began adding the code necessary to load an object file. This would be embedded inside a PE file, for which I have not done anything to parse yet.
Found a Microsoft document on how code signing is implemented in PE files.